Add limit of 65535 shoot types

+++ src/combat.h
@@ -69,7 +69,7 @@ struct CombatParams {
 	CombatOrigin origin = ORIGIN_MOVE;
 
 	uint8_t impactEffect = CONST_ME_NONE;
-	uint8_t distanceEffect = CONST_ANI_NONE;
+	uint16_t distanceEffect = CONST_ANI_NONE;
 
 	bool blockedByArmor = false;
 	bool blockedByShield = false;
@@ -267,7 +267,7 @@ class Combat
 		static ReturnValue canDoCombat(Creature* attacker, Creature* target);
 		static void postCombatEffects(Creature* caster, const Position& pos, const CombatParams& params);
 
-		static void addDistanceEffect(Creature* caster, const Position& fromPos, const Position& toPos, uint8_t effect);
+		static void addDistanceEffect(Creature* caster, const Position& fromPos, const Position& toPos, uint16_t effect);
 
 		void doCombat(Creature* caster, Creature* target) const;
 		void doCombat(Creature* caster, const Position& position) const;
